refactor: 폴더구조 역할별->도메인별 변경
