+++ server/src/purchase-item/purchase-item.types.ts
@@ -1,5 +1,3 @@
-import { Category } from "@prisma/client";
-
 export const DATE_TYPES = ["purchased", "created"] as const;
 export type DateType = (typeof DATE_TYPES)[number];
 
@@ -25,22 +23,4 @@ export interface GetPurchaseItemsQuery {
   backorderOnly: boolean;
   sortBy: PurchaseItemSortBy;
   sortOrder: SortOrder;
-}
-
-export interface CreatePurchaseItemDto {
-  productName: string;
-  category: Category;
-  color?: string;
-  size?: string;
-  extraOption?: string;
-  unitPrice: number;
-  quantity: number;
-  backorderQuantity?: number;
-}
-
-export interface CreatePurchaseDto {
-  vendorName: string;
-  purchasedDate: string;
-  receipt: string;
-  items: CreatePurchaseItemDto[];
 }
